Fix files for window exe creation

See merge request technology-factory/grasp-sensor/discovery/rcdiscover!6

+++ rcdiscover/discover.cc
@@ -137,7 +137,7 @@ bool Discover::getResponse(std::vector<DeviceInfo> &info,
       {
         count--;
 
-        if (select(sock+1, &fds, NULL, NULL, &tv) > 0)
+        if (select((int)(sock+1), &fds, NULL, NULL, &tv) > 0)
         {
           // get package
